feat: 클라이언트 컴포넌트로 분리하여 패치하기
